SW-4693 Add extra permanent cluster count
To support mandatory permanent clusters in newly-added areas when planting zones
are expanded, add a new value to the zone information to hold the number of extra
permanent clusters that should be included in observations.

The new value is included in the admin UI and is incorporated into the cluster
count calculation in the UI, but isn't directly used by the server-side
application code yet.