SW-4716 Test plot positions in ObservationServiceTest (#1627)
In preparation for changing how monitoring plot allocation works, change the
assertions in the monitoring plot allocation test case in ObservationServiceTest
so they check for valid geometric relationships between the plots and the
subzones rather than checking for specific plot IDs.

The setup part of the test will need to change when the plot allocation code
is revamped, but the assertions should be able to remain untouched, such that we
can verify that the same requirements are being met before and after the change.

To make geometry-based tests like this one easier to read and write, add
convenience arguments to the planting site/zone/subzone/monitoring plot insert
functions in DatabaseTest to allow boundaries to be defined in terms of origin
point and width/height.