SW-4731 Create missing permanent clusters (#1645)
`PlantingSiteStore.ensurePermanentClustersExist()` will, as its name suggests,
ensure that the required number of permanent clusters exist in each of a site's
planting zones, creating new monitoring plots or adding existing temporary plots
to clusters as needed.

For existing planting sites, this is a no-op because a full set of permanent
clusters is created at shapefile import time and this function thus has no work
to do, but future updates will get rid of the import-time plot creation and will
require this logic to be in place.