Fix facility creation race condition
In end-to-end tests, we were sometimes sending two facility creation requests at
the exact same time from different browsers. This exposed a race in the facility
creation code where the same facility number would be used for both facilities,
causing one of the creations to fail with a unique key violation.

Fix it by serializing facility creation using a row lock on the organization.