Add light/dark toggle to admin UI map viewer
Once monitoring plots are created on demand, there will no longer be a grid of
them covering the entire planting site; instead they will be scattered around an
otherwise empty map. For the kinds of very large planting sites we sometimes
create during testing, this makes them hard to see on a fully zoomed-out map.

Add a "Light/Dark" button to the admin UI's planting site map viewer that toggles
between the default satellite-photo map and a dark one.

In addition, for temporary and permanent plots, put the plot borders in their own
map layers that are rendered as outlines rather than using the "fill" layer's
outline color option. On zoomed-out maps, Mapbox completely hides small polygons
that are rendered as "fill" layers, but renders "line" layers as single pixels;
this makes it possible to see where the plots are.