Reduce clutter when calling geometry test helpers
Update the helper functions that are used to construct geometries in tests so they
accept any kind of number rather than only `Double`. We almost always call them
with whole-number arguments anyway.

The `point()` helper method defaults to using the X coordinate as the Y
coordinate, but we were passing it the same value twice in a lot of places. Remove
the redundant second argument in those cases.

There was an alternate helper function to create `Point` objects, but it was only
used in one test; remove it and extend `point()` to support its functionality.