SW-4802 Move findUnusedSquare to helper class
For user-defined planting zone boundaries, we'll want to validate that the
boundaries are large enough to hold monitoring plots before inserting the zone
data into the database.

Ideally, we wanto to do that using the same "find a plot location" algorithm
that's used in real observations. That code is currently in `PlantingZoneModel`
which can only represent planting zones that exist in the database.

Move the zone search logic into a helper class `UnusedSquareFinder`.

This is mostly a straight copy-paste of the existing implementation, with the
following changes:

- The logic for excluding existing monitoring plots now lives in
  `findUnusedSquares` since it was only ever used via that function anyway.
  This requires changing the unit test for that functionality.
- The various functions that used to be defined inside `findUnusedSquare` are
  now just regular methods in the new class.
- Likewise, some of the local variables from `findUnusedSquare` are now
  instance variables in the new class.
- It's no longer possible to specify a grid interval thats different than the
  square size. We weren't ever doing that anyway.