SW-4865 Store batch ID on new accession withdrawal
When seeds are withdrawn to nurseries, we want to store the batch ID that's
created by the withdrawal. The ID was being passed to `WithdrawalStore` but it
wasn't actually saving the value to the database.

Fix the bug and rerun the migration that backfills batch IDs in any nursery
transfers that happened since the initial run.