Accept Polygon planting site boundaries
Allow clients to supply planting site boundaries as either MultiPolygon or Polygon
geometries. This will save clients the trouble of converting single-polygon site
boundaries to MultiPolygon solely because it's required by the API.

This requires updating our OpenAPI schema customization code to handle properties
that are references to payload classes but have explicit annotations to list which
types are allowed.