SW-4911 Create schema for accelerator tables
We'll shortly be creating a number of new tables for accelerator workflow data,
and there will be enough of them that they'll clutter up the `public` schema
significantly. Create a new `accelerator` schema and move the existing
accelerator-related tables there.

This is purely for the sake of keeping the data model better organized; there's
no change in behavior here.