Fix CohortStoreTest (#1751)
The test was checking for a specific ID being created but wasn't resetting the
ID sequence on the table, so if another earlier test inserted into the same table,
this test would fail.

We don't actually care which specific ID is used, so fix the test by ignoring the
ID entirely when asserting that the expected values were inserted into the
database.