SW-5022 Allow internal users to read deliverables
The permission checks for reading deliverables and submissions required the user
to be a member of the organization whose projects were being queried, and also
allowed contributor users to read deliverable data.

Add permissions for deliverable reading that allow access by internal users with
the appropriate global roles but not organization users with contributor roles.