Make EnumsTest output more useful
When EnumsTest detects that English-name entries are missing from a properties
file, make its assertion failure message include a list of lines that can be
copied into the properties file as a starting point.

The English strings will still require human review, and, if needed, comments
will still need to be written to clarify terminology for translators. But this
should make it slightly less annoying to add missing enums to the strings files.