SW-5070 Add project accelerator details to schema
Create new tables for projects' accelerator-specific details.

Add a country code column to the existing projects table; initially it won't be
exposed to end users via the API but that may change in the future.

The list of land use model types for a project, likewise, lives in the public
schema rather than the accelerator schema, since it's also information we may
support for non-accelerator projects in the future.