SW-5089 Add endpoint to list accelerator orgs
`GET /api/v1/accelerator/organizations` returns a list of organizations with the
Accelerator internal tag and a list of each of those organizations' projects.
By default, it omits projects that are already assigned to participants.

Access depends on the user's ability to read internal tags, which is extended to
include TF Expert and Read Only global roles.