SW-5089 Global roles grant limited read access (#1790)
Allow users with the Super-Admin global role to read all organizations and all
projects, and allow users with the other global roles to read organizations with
the Accelerator internal tag and their projects.

Currently, this does not allow read access via the search API, only via the
dedicated GET endpoints for organizations and projects. It also doesn't grant
read access to other entities such as batches or accessions, just the basic
data (name, etc.) of the organizations and projects.