SW-5089 Global roles allow searching accelerator orgs (#1797)
Allow users with appropriate global roles to read organization and project
data for organizations with the Accelerator internal tag via the search API.

This changes the behavior of the search API for privileged users! Previously,
it would only return results from the user's organizations, but now it can
also return results for other accelerator organizations. This won't affect
any searches that explicitly filter on project or organization ID, but will
affect unconstrained searches.