SW-4732 Create missing temporary plots (#1824)
When an observation starts and temporary plot boundaries are selected on the site
map, create new monitoring plots in the database for any boundaries that aren't
already there.

For existing planting sites, this is a no-op because all possible plot boundaries
are inserted into the database at shapefile import time. But future updates will
get rid of the import-time plot creation to better support editing maps after site
creation.

This is analogous to commit 14d90e78c which added on-demand cration of permanent
clusters.