SW-4796 Reduce repetition in ObservationServiceTest
Add a helper method to insert clusters of monitoring plots.

In addition to reducing repetition in the test code, the helper method also places
the plots in the correct geographic positions. This is a step toward updating the
test to continue to pass when the plot replacement logic is updated to support
on-the-fly creation of plots based on the site map.