Fix duplicate projects on voters admin page
If a project had vote decision records for multiple phases, we were listing the
project multiple times on the voter management page in the admin UI, but only
showing the voters for the current phase.

Change the database query and the model class to be aware that there can be
multiple vote decisions for a project, and make the admin UI only show the current
phase.