SW-5188 Implement undo of nursery withdrawals
Add an `undoWithdrawal` method to `BatchStore` that reverses the effects of a
previous withdrawal on seedling batches.

Take undo withdrawals into account in the nursery stats calculations.

Currently, there is no way for an end user to use this code; it will be hooked
up to the API in a later change.