SW-5187 Add valid user-supplied withdrawal purposes (#1841)
In preparation for adding a new withdrawal purpose for undoing previous
withdrawals, update the "create withdrawal" endpoint to use its own list of
valid withdrawal purposes.

When the new purpose is added, it will appear in the `WithdrawalPurpose`
enum, but with this change it won't be listed as a valid user-supplied
purpose for withdrawal creation in the OpenAPI schema.