SW-5188 Database changes to support withdrawal undo (#1842)
Update the database schema to support a new nursery withdrawal purpose of "Undo"
that, as the name suggests, is an undo of a previous withdrawal. An undo
withdrawal has negative withdrawn quantities and includes a reference to the
original withdrawal it undoes.

There's no behavior change here, but this does affect the API in that "Undo"
is now a valid withdrawal purpose.