SW-5187 Add undo dates to search API (#1856)
In the web app, we need to show the date of the original withdrawal as part of the
withdrawal log row for an undo operation. Currently, the client has to request the
dates from the server after getting the initial set of search results. Add fields
for the dates on both sides of an undo operation to avoid the extra API calls.