SW-4796 Generalize plot replacement tests (#1861)
In preparation for making monitoring plot replacement create new plots on the fly,
update the tests to focus on the desired properties of the end result rather than
on the details of the existing implementation.

The existing implementation involves swapping permanent cluster numbers such that
a given cluster number refers to different plots after the replacement. We only
actually care that the original cluster no longer has the number, and that some
other cluster does, not that the two clusters' numbers were switched.

The updated tests continue to pass with the existing implementation; a subsequent
change will update the implementation without touching the tests to ensure that
existing functionality hasn't been broken.