SW-4733 Don't generate all plots on shapefile import (#1863)
Currently, when a shapefile is imported to create a new planting site, or when a
new planting site is created via the admin UI, we create a grid of all the
possible monitoring plots across the entire site. For large sites, this can be
a bit slow and also fills the monitoring plots table with plots that won't be used
any time soon, if ever.

Now that the application code can create new plots on the fly as needed, there is
no longer a reason to create all the plots up front. Remove that part of the
import process.