SW-4736 Delete unused monitoring plots (#1865)
Now that we create monitoring plots on demand and no longer create them as part of
shapefile import, existing planting sites have a full set of all potential
monitoring plots and new ones won't.

For consistency, get rid of all the plots in existing sites that weren't included
in any observations. This will leave us with a small number of plots which should
improve query efficiency, especially for large sites.