SW-5179 Created Atlassian Client for support tickets (#1838)
JIRA is no longer supporting a client library so we implemented a thin HTTP layer using ktor to submit REST APIs.

An AtlassianResource interface is used to build the request including path, header and request body and each endpoint will implement this interface. This is to maintain strong typing and ensure that the payload/request are correct.

The tests should demonstrate the use cases of this client.