SW-5264 Populate planting site history
When a planting site with map data is created, or a simple planting site with a
map is updated, record the map geometries in `planting_site_histories` and
the child tables for zones and subzones.

The server doesn't currently support updates of detailed planting site maps,
so zone and subzone history records are only inserted at creation time; updating
of detailed planting site maps will be added in upcoming changes.