Parameterize planting site model IDs
In preparation for adding logic to edit existing planting sites, introduce
"new" and "existing" variants of `PlantingSiteModel` similar to the pattern we
use on other model classes.

The wrinkle here is that the parameterized nullability needs to also apply to
the models for planting zones and planting subzones.

This will allow us to use the model classes to represent a site/zone/subzone
hierarchy where some of the objects haven't been inserted into the database
yet.

This is mostly replacing `PlantingSiteModel` with `ExistingPlantingSiteModel`
in the application code (same with zone and subzone models), but also required
replacing a secondary constructor with a function so as to specify its return
type.