SW-5228 Use model to create new planting sites (#1887)
`PlantingSiteStore.createPlantingSite` now takes a planting site model instead
of a set of parameters for individual fields. (Planting seasons are still a
separate parameter for now.)

This is in preparation for moving the creation of planting zones and subzones
from `PlantingSiteImporter` to `PlantingSiteStore`, which will help support
user-drawn zone and subzone boundaries (where there is no shapefile to import)
as well as giving us a consistent place to create the initial map history
records for zones and subzones.

Aside from a new convenience function for creating `NewPlantingSiteModel`
objects, there is no functional change here, just moving values from function
parameters to model fields.