SW-5230 Remove shapefile validation options (#1889)
Early on in the development of the tracking features, all our test shapefiles were
invalid in one way or another, so we needed to be able to disable the various
validation checks in order to test the code.

Now that we have shapefiles whose geometries are valid and whose only "problem" is
that they sometimes lack target planting densities for zones, there's no need to
continue supporting invalid files. Remove the concept of "validation options" from
the importer and start treating target planting density as an optional value with
a default.