SW-5230 Convert shapefile coordinates to long/lat
We always use the WGS 84 (longitude/latitude) coordinate system when storing
planting site geometries in the database. Currently, that conversion is done at
write time using PostGIS's coordinate transformation functions; our
`GeometryBinding` class wraps geometry values in the appropriate functions when
it generates SQL for write operations.

The downside of this is that it means when we are doing geometry calculations
on a shapefile at import time, we're potentially using a different coordinate
system than what will end up in the database. That's not a problem with the
existing write-once model, but when we support updating maps, it'll make it
more difficult to compare the old and new maps.

Update the shapefile importer to do the coordinate system transformation in the
application code before doing its geometry calculations.

The coordinate system transformation involves floating-point calculations that
can introduce slight discrepancies in boundaries; this isn't noticeable on
actual maps but caused failures in tests that were checking for strict overlap
of geometries. Change those tests to allow a slight fuzz factor like we already
allow in the application code.