SW-5228 Move planting site validation to model classes (#1904)
Currently, `PlantingSiteImporter` does two kinds of validation: checking that the
shapefiles are correctly formed (correct attribute names, etc.) and checking
that the maps in the shapefile obey our constraints on planting sites.

Move the constraint checks to `PlantingSiteModel` and `PlantingZoneModel` so we
can validate planting site maps that aren't in the form of shapefiles. This is
a step toward unifying shapefile-based and API-based detailed planting site
creation and updating, though for now, the validation is still only done during
shapefile import.