Move documentation to docs directory (#1925)
The top-level directory was getting a little cluttered with documentation files,
and we're going to be adding more of them as we merge the document producer into
Terraware. Move them to the `docs` directory and make them viewable in GitHub
Pages.

This is mostly just straight renames, except that part of `README.md` stays at
the top level and the rest of it is renamed to `docs/SETUP.md`.