Add DSL for constructing test planting sites
To support tests that need to construct planting site maps in specific
configurations, add a builder class that acts as a planting site DSL.
This allows site maps to be specified succinctly, which makes the test cases
easier to follow.

This is mainly motivated by the upcoming support for planting site map
editing, the tests for which require constructing lots of site maps. But it
also applies to some existing tests. `PlantingSiteModelTest` is updated to
use it.