SW-5228 Calculate changes to site maps (#1953)
* SW-5228 Calculate changes to site maps

In preparation for allowing editing of planting site maps, add logic to
calculate the differences between an existing planting site and a site map
requested by the user, and to detect edits that are disallowed according to the
product requirements.

The internal representation of a site edit will be used as the basis of both the
confirmation step in the UI ("this edit would do X, Y, and Z") and as input to
an update function that will apply the requested edits.

The code in this change only concerns itself with site geometry, not with any
of the other edit behaviors related to active observations or the presence of
planting data. Upcoming changes will cover those requirements.

* Change order of assertHasProblem arguments to allow adding optional arguments more cleanly

* Add test cases for zone/subzone deletion, exclusion area reduction; fix bugs uncovered by the new tests