SW-5230 Ignore exclusions in site/zone/subzone areas
In the upcoming changes to support planting site editing, we're supposed to update
the site/zone/subzone area in hectares if an exclusion grows or shrinks. But we
were disregarding exclusions in the initial area calculations at site creation
time, which would have been inconsistent with the new behavior.

Internally, areas were being calculated as part of writing the site data to the
database, and the site/zone/subzone models were being intentionally constructed
with dummy values for their areas, meaning that newly-constructed models weren't
representative of the actual site data.

Move the area calculations into the model classes (or rather, the `create`
functions we use to construct new models) and make them take exclusions into
account.