Set and use correct factory on geometry objects (#1977)
JTS Geometry objects remember which factory created them, so we can reuse that
factory when we're creating derived geometries.

But when we were reading geometry values from the database, we were using a
geometry factory that didn't have the correct SRID to indicate that the
coordinates were longitude and latitude, so reusing those objects' factories
would produce geometries with incorrect SRIDs.

Set the correct SRID when constructing new geometries, and reuse the existing
factory when deriving new geometries from existing ones.