SW-5230 Add shapefile update to admin UI
Add an upload form to the planting site details page in the admin UI to allow
admins to upload revised shapefiles for existing detailed planting sites.

By default, it does a dry run and reports the change in plantable area of the
site as a whole and of each zone. The admin can then choose to actually apply
the changes to the site.