SW-5431 Don't require site and zone shapefiles
The GIS team currently produces the zone and site shapefiles by merging the
subzone boundaries. There's no need for them to do that manually; we can do
geometry merging as part of the shapefile import process.

Update the importer so that it expects a zipfile with either one or two
shapefiles: the subzones and optionally the exclusions.

Zone-level settings such as target planting density are pulled from the
subzones shapefile; they only need to be set on one subzone per zone.