Stop using deprecated Jackson API
Our custom deserializer for geometry objects throws an exception when a GeoJSON
value has an unexpected or missing geometry type. The exception includes the
location in the JSON object where the failure happened.

We were using an old Jackson API method to retrieve the location. In the next
Jackson version, that method is marked as deprecated, which will cause a build
failure when we upgrade. Switch to the newer version of the method.