Fix code docs generation (#2006)
The latest Spring Boot version uses a Jackson version that's not compatible with
Kotlin's documentation generator tool (Dokka). Override the Jackson version when
generating docs, per the suggestion in the Dokka bug report about the problem.

If/when Dokka is updated to work with more recent Jackson versions, this change
can be reverted.