Make plot-sized boundaries in tests
Currently the width and height parameters for `insertPlantingSite()` and
friends just add the values to the raw coordinates in the default coordinate
system, meaning a site with a width of 5 actually has a width of 5 degrees
longitude.

Instead treat them as multiples of the monitoring plot size, which will make it
more convenient to construct test data for planting site editing tests.