Add helper for inserting test planting sites
A bunch of the planting site tests have a repetitive pattern of inserting a series
of rows into the database to get a planting site with plants delivered to a
subzone. That pattern was encapsulated in a function in `ObservationStoreTest`;
move it to a helper class and use it in `ObservationServiceTest` too.

The new helper class also has the plant totals assertion functions from
`ObservationStoreTest` for use in future tests of other classes.