SW-5472 Allow plot replacement in non-completed subzones
When we're starting an observation, any subzone that has reported plants is
eligible to have monitoring plots. We use the same criterion when replacing a
monitoring plot due to a site map edit.

But when the user requested a plot replacement, we were instead looking at
whether the subzone was marked as having _completed_ planting.

Update the plot replacement code to make it behave consistently with plot
selection in other contexts.