Object representation of map validation failures
Previously, map validation failures were represented as English strings for
display in the admin UI. But in the API, we'll need to turn them into
machine-readable values.

Add validation failure types for all the validation checks in `PlantingSiteModel`
and `PlantingZoneModel`, and change those classes to return a list of failure
objects rather than a list of strings.

Shapefile validation failures are still represented as English strings since
shapefile upload is going to remain an admin UI function. But to make the
distinction between map validation and shapefile validation clearer, the
exception that's thrown for shapefile problems is renamed to
`ShapefilesInvalidException`.