Move site validation from controller to model
`PlantingSitesController` validates that geometries are of acceptable types
since the request is syntactically invalid and can't be turned into a
`PlantingSiteModel` otherwise.

It was also doing a little structural validation of the site, but those validation
checks had nothing to do with whether the payload was well-formed. Move those
checks to `PlantingSiteModel`.