SW-5516 Add new permanent clusters to observation
When area is added to a planting site with an observation in progress, we need to
create new permanent clusters in the newly-added area and include them in the
in-progress observation.

We were creating them, but not adding them to the observation because they were
created as part of applying the map edit; the code that updates the observation
plots was only adding plots that needed to be created to replace previous ones
because of a zone shrinking.