SW-5566 Add variable owners
Allow variables to be assigned to owners on a per-project basis.

Initially, this will be used to assign numbered sections to owners during the
feasibility study authoring process, but this implementation is more general
and allows any variable to have an owner.

Variable ownership is not considered part of the variable's value; it's a
separate piece of metadata that can exist even for variables that don't yet
have values in a project.

In anticipation of changing the variables API to stop requiring documents
(which will be necessary to support questionnaire deliverables), the new
endpoints added here use project IDs rather than document IDs.