SW-5582 Project-based variable value endpoints
To support variable values that aren't associated with documents, add versions of
the variable value endpoints that take a project ID rather than a document ID in
the URL path. The existing document-based endpoints continue to work as before;
once the web app has been updated to move to the project-based ones, we can remove
the document-based versions.