SW-5499 Return workflow details with variable values
Return the current workflow details for variables as part of the "list values"
response. Internal comments are only returned if the user has permission to see
them (that is, if they have an appropriate global role).

In addition to the new payload fields, there is a behavior change here: the list
of values for a variable can be empty if the variable has workflow details in the
project but doesn't have any values yet. Previously, the values list for a
variable was guaranteed to be non-empty.