SW-5565 Remove sensor kit endpoints
We are no longer supporting sensor kits in the Terraware UI, so remove the code
related to setting them up from the web app and querying their data.

Since we are still using a sensor kit for in-house monitoring, the endpoints
that existing sensor kits use to report data remain in place, as does the
admin UI functionality for editing device data.