Add permissions for internal workflow details (#2099)
Variable owners are just one of the internal-use-only workflow-related details
of a variable in a project. Replace the read and update permission on project
variable owners with a broader permission for internal-only project variable
details.

No difference in behavior here; upcoming changes will use the new permissions
in other contexts.