SW-5499 Add variable workflow details table (#2100)
Add a table to hold workflow details (internal comment, feedback, status) for
variables. The new table actually stores the _history_ of the workflow details
since we want to keep track of edits over time. Historical values won't be
exposed via the API initially but will be stored for later use.