Add alpha-3 codes, eligibility to countries table
The accelerator team wants to use 3-letter ISO country codes instead of 2-letter
ones to track accelerator applications.

Since we use alpha-2 codes internally, add a column to the countries table for
the alpha-3 codes so we can easily map between the two.

Also add an "eligible" flag to indicate which countries are eligible for the
accelerator program.