SW-5605 Stop using hardwired facility IDs in tests
As a step toward being able to run tests in parallel, eliminate nearly all uses of
hardwired facility IDs in tests, instead letting the database allocate IDs
dynamically.

The "insert" methods in `DatabaseBackedTest` whose facility ID arguments used to
default to the hardwired ID now use the last inserted ID instead; this eliminates
the need to pass in facility IDs explicitly in a number of tests.

The exception is `PermissionTest` where the numeric values of the IDs affect the
test logic; that one will require more extensive reworking.