Fix classpath conflict (#2170)
The access log library we're using depends on an older version of some of the
classes for the Tomcat embedded web server, which can cause the server to fail to
start up in dev environments if Gradle decides to put that library before Spring
Boot in the classpath.

The failure is due to https://github.com/gradle/gradle/issues/20286 and can happen
intermittently.