Split "Waitlist" status into 3 internal ones
The PRD was updated to replace the "Waitlist" application status with three
internal-facing statuses that all map to "Waitlist" for external users.

This change replaces the existing "Not Accepted" status with one of the new
ones, which should be safe since there shouldn't be any actual applications
with that status yet given that we haven't written the console code to let
people edit statuses yet.