SW-5662 Implement non-variable pre-screen checks
Change the response of the "submit application" API endpoint to return a list of
problems and an updated copy of the application data, including its updated
status.

Implement the check logic for all the pre-screen requirements that don't involve
fetching values from variables; the variable-based ones will require the
variables to be defined which isn't done yet. But these initial checks should be
enough to support testing the submission logic in the frontend.