Exclude springdoc-openapi from Renovate group
We can't upgrade springdoc-openapi until a particular bug is fixed. Renovate
is supposed to not suggest upgrading dependencies to versions it has previously
suggested, but that only works for dependencies that aren't grouped together.

Add an exclusion rule to stop springdoc-openapi from being grouped with the
other Spring dependencies so it won't continue to be included when other
dependencies in the group have new versions.