SW-5737 Improve robustness of shapefile import
We recently got a complex shapefile that failed to import for a couple reasons.

The big one was that it used 4-dimensional coordinates (the fourth being a
non-spatial "M" value that's used to hold application-defined values) which caused
some of the geometry calculations to bomb out.

It also had a lot of complex borders between subzones, which, thanks to
floating-point precision limitations, caused the calculated zone and site
boundaries to have small holes in them.

Make the importer more robust to this kind of shapefile:

- Strip the Z and M dimensions, if present
- Use fixed-precision coordinates to avoid floating-point errors
- Remove small gaps between adjacent subzones when calculating zone and site
  boundaries