SW-5628 Revert change to default Student's t value
Set the default value for the "Student's t" input to the calculation of the number
of monitoring plots per zone back to its old value of 1.645 to get a 90%
confidence interval. This causes the default number of clusters and plots to
increase, which also increases the number of plots and clusters that are added to
new areas when a zone's size is increased.