SW-5741 Always set application internal name (#2219)
When an application is created, set its internal name using "XXX" as the country
code, updating it the first time a country can be determined from a boundary.

The internal name continues to be nullable in the API because it isn't supposed
to be visible to external users. Unfortunately, we weren't actually enforcing that
visibility rule; start doing so.