SW-5672 Remove the requirement of an active deliverable when managing participant project species (#2223)
Previously, due to requirements for notifications related to management of participant project species, the need for an "active deliverable" was present. An "active deliverable" is a deliverable that belongs to a module that has a cohort module related to the participant project in question that is occurring "now". Since we want to allow participants to interact with their participant project species lists at any time, the requirement was removed.

If there is an active deliverable, it will still be used to send out notifications, and we still ensure that a submission for the project exists against the deliverable.

If there is no active deliverable, we attempt to find the most recent deliverable in the past, and use that to send out notifications and ensure the submission.

If there are somehow no deliverables related to the participant project in question, we still allow the operations to proceed, effectively allowing the participant project species list to exist outside the context of a deliverable.

In the event there are multiple species list deliverables for a related cohort module (this is not likely to be the case in practice, but is allowed by the system), the one with the highest ID is used.